# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations/current/users": {
      "get": {
        "operationId": "OrganizationUsersController_listOrganizationUsers",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "suspended", "archived"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A filtered, cursor-paginated list of users in the caller's own organization, ordered by name then id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationUserListResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed. Error code: VALIDATION_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed."
                }
              }
            }
          },
          "401": {
            "description": "No valid, active session was supplied. Error code: AUTHENTICATION_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "No valid, active session was supplied."
                }
              }
            }
          },
          "403": {
            "description": "The authenticated account is not active. Error code: ACCESS_DENIED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "ACCESS_DENIED",
                  "message": "The authenticated account is not active."
                }
              }
            }
          },
          "500": {
            "description": "An unexpected internal error occurred. Error code: INTERNAL_SERVER_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 500,
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "Internal server error."
                }
              }
            }
          }
        },
        "summary": "Search the authenticated user's organization directory",
        "tags": ["organizations"]
      }
    }
  },
  "info": {
    "title": "AtlasHQ API",
    "description": "AtlasHQ REST API contract. Operation IDs use Controller_method.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z0-9_]+$"
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "details": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "minItems": 1,
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 1
                      },
                      {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      }
                    ]
                  }
                },
                "message": {
                  "type": "string",
                  "minLength": 1
                },
                "code": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": ["path", "message", "code"],
              "additionalProperties": false
            }
          },
          "correlationId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["statusCode", "code", "message"],
        "additionalProperties": false
      },
...
      "OrganizationUserListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "email": {
                  "type": "string",
                  "format": "email",
                  "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                },
                "status": {
                  "type": "string",
                  "enum": ["active", "suspended", "archived"]
                },
                "organizationRole": {
                  "type": "string",
                  "enum": ["admin", "member"]
                }
              },
              "required": ["id", "name", "email", "status", "organizationRole"],
              "additionalProperties": false
            }
          },
          "pageInfo": {
            "type": "object",
            "properties": {
              "limit": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100
              },
              "nextCursor": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 1
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "hasMore": {
                "type": "boolean"
              },
              "total": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": ["limit", "nextCursor", "hasMore"],
            "additionalProperties": false
          }
        },
        "required": ["items", "pageInfo"],
        "additionalProperties": false
      }
    }
  }
}